Allow for multiple Providers
This allows you to have a concept of inner and outer contexts.

This is valuable for situations where you have an admin (or similar)
user who is viewing the page as a user (or similar) should see it. This
allows you to have a "current user" context and any number of "user I'm
looking at" contexts.

How does it work?

- Initialize your own React Context and `use` hook.
- Pass the Context into `<PrefabProvider>` as `ContextToUse`
- Use your `use` hook to access that context
- Continue using the `usePrefab` hook to access the original
  `<PrefabProvider>`'s context

See the tests for an example.

Temporarily disable prettier in CI since it is giving false positives.